requirements.txt
annotated-types==0.6.0
anyio==3.7.1
black==23.12.1
click==8.1.7
fastapi==0.105.0
flake8==6.1.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
load-dotenv==0.1.0
mccabe==0.7.0
mypy-extensions==1.0.0
mysql-connector-python==8.2.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
protobuf==4.21.12
py_mysql_connector==0.1
pycodestyle==2.11.1
pydantic==2.5.3
pydantic_core==2.14.6
pyflakes==3.1.0
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.25.0
